o when processing subscription update events, always activate the send independent of whether receive bytes are pending or not. This improves event latency, and allows events to flow if we are blocked due to the service postponing an IO operation. This was a bug fix. o when finishing processing input activate a send if more than one half of the TCP buffering would be used. In the past the send was not activated until either the send buffer was full or the receive buffer (including any bytes waiting in the socket) was empty. This could cause latency and performance issues because send receive piplining would not occur if they have very large buffers due to a large EPICS_CA_MAX_ARRAY_SIZE. o always activate a send if we enter a service is postponing IO state. Users would probably prefer to receive past responses and not wait until the postponed IO initiates.
o casStrmClient::inBufBytesAvailable renamed to
casStrmClient:: inBufBytesPending reflecting the fact that we nlonger use
a socket ioctl to check how many bytes are pending in the sockets
input queue
o casStrmClient::outBufBytesPresent renamed to
casStrmClient::outBufBytesPending for consistency with inBufBytesPending
o removed the eventFlush function
o removed sendBlocked flag
o call processMessage directly from the IO completion callback
that restarts when in an IO postponement state. This makes the logic easier
to understand and maintain, but isnt a functional change
o call processMessage directly from the receive callback.
This makes the logic easier to understand and maintain, but isnt
a functional change
o call processMessage directly from the send callback.
This makes the logic easier to understand and maintain, but isnt
a functional change