dbtools: Added dbltExpand.c
This is a test program which implements the template expansion
capabilities of msi, but using the dbLoadTemplate.c code.  It should
be useful for testing, to make sure that both versions generate the
same output.  It won't work on Windows at the moment due to a clash in
the function decorations of the dbLoadRecords() function.