posix/osdThread: validate pcommonAttr before using
This addresses a bug when a non-EPICS thread would create
an epics threadInfo context via 'createImplicit()'. The
computed OSI priority was wrong if EPICS was executing with
SCHED_FIFO enabled but the non-EPICS thread used SCHED_OTHER.

This patch does some cleanup and makes sure that
 - pcommonAttr->valid (i.e., the SCHED_FIFO could
   be used for determining its min/max priorities).
 - pcommonAttr->policy == threadInfo->policy
when performing an operation on threadInfo that
involves pcommonAttr.

A new test case 'nonEpicsThreadPriorityTest' has been added
which verifies that a non-EPICS/SCHED_OTHER thread gets the
correct osiPriority (0) assigned when EPICS runs under
SCHED_FIFO. The caveat, however, is that this test only works
when executed with sufficient privilege (otherwise the test
in question will be skipped).