epicsShareFunc void epicsAssert(const char *pFile, const unsigned line,     const char *pExp, const char *pAuthorName);
is used by a small number of files in a release build, but the definition is only exported - by the header - for a debug build.
(most files use assert, which is #defined to nothing in a release build.)

The function is (correctly) defined for both debug and release builds in osdAssert.c.
So there's no reason for it not to be exported by epicsAssert.h