Resolve some merge issues.
There are changes that made it through the merge, but I think are no
longer needed.

osdNTPGet doesn't appear to be used anywhere, I think its use was
removed but the definition of the function made it through the merge.

The include of sysinit.h existed on one branch, but not on another and
it caused a compile error, so I removed it for now.

rtems_bsd_set_vprintf_handler failed on RTEMS 4.10.2 so was commented out.