IPv6: Use aToIPAddr in aToIPAddr46 for non-`[]`
Use the "old" aToIPAddr unless we have a IP address surronded by []
This allows the old parsing to be use (IPV4 address as long).
It will use the IPv4 address of a host that has DNS entries for
both IPV4 and IPv6.
So we are slightly more backward-compatible.

Note: the getaddrinfo() may be replaced later, not all platforms seem
to have it.