IPv6: printing a socket: it is not an int
When printing a socket under windows, we get a warning:
a SOCKET is a "long long unsigned" (under Windows)

Cast it into a an "int" when printing.
On most platform it is already an "int", under Windows we don't print
the whole SOCKET.
Using a "long" does not help, since a "long" under Windows is 32 bits,
the same as an "int".