IPv6: udpiiu.cpp change call to epicsSocket46optIPv6MultiCast()
This may be a hack, further invstigations are needed.
When searching for a PV, we need to call epicsSocket46optIPv6MultiCast()
on the socket that sends out the UDP search message.

Setting this option epicsSocket46optIPv6MultiCast() for different scope_id
works on Linux, but not on MacOS.

Solution:
Set the epicsSocket46optIPv6MultiCast() before sending
and especially: before receiving.

A better solution may be to use a different socket for each IPv6 scope_id.
This will need more changes for recvfrom(): a select() or poll() is needed
when trying to recvfrom() on different sockets