IPv6: Make global beaconSocket local
First step to have more than one socket for the beacons:
Make the global beaconSocket local to caservertask, and hand it over to
online_notify by in a new-defined config struct.

Note: The "one socket" needs probably to become a list of sockets.
This will be done in the next commit.