IPv6: Remove EPICS_HAS_IPV6_LINUX
Unify the socket handling for Linux and non-Linux systems:
IPv4 is send on a socket created with AF_INET, IPv6 with AF_INET6

The Linux kernel seems to be able to handle a mix of both on the
same socket very well, but e.g. MacOS does not.
And probably other BSD-ish systems do do that either.