IPv6: casw/repeater: Use an IPv4 socket
casw diddn't work due to a mess-up between IPv4 and IPv6 addresses.
Revert the IPv6 communication between casw and the repeater:
The repeater does not handle IPv6 yet (no way to embedd the received
IPv6 address in a fanout-ed packet).
So let's do casw and repeater in IPv4 only for the moment.

Note:
There seem to be 2 problems to be solved:
- The beacon message(s) CA_PROTO_RSRV_IS_UP is too short to transport IPv6
  addresses
- The repeater itself need to use multiple sockets for IPv6 multicast,
  similar to what we have in udpiiu::udpiiu today