IPv6: Preserve errno in osiSock.c when printing
The repeater relies on errno being the one from the last socket
operation.

See makeSocket():
if bind() fails, status is set to SOCKERRNO.
"SOCKERRNO" is "errno" on all systems except Windows.

In epicsSocket46Bind() the socket error is printed as well,
which seems to clobber errno.
Save & and restore it around all debug-printing.

    if ( sockerrno == SOCK_EADDRINUSE ) {