IPv6: epicsSocket46Connect() needs sockets_family
When working in a mixed environment, IPv4 caget/monitor/info did not
work anymore.
The reason is are these commits:
  IPv6: Do not use osiSockIPv4toIPv6() in epicsSocket46Connect()
  IPv6: Remove osiSockIPv4toIPv6()

The reason for those was mainly the repeater:
registration messages are still send via IPv4 to localhost.
And no IPv4->IPv6 mapping was needed.

However, cainfo and friends use an IPv6 socket to connect to IPv4 host,
and here we need osiSockIPv4toIPv6().
In order to distinguis those 2, add the family of the socket
to epicsSocket46Connect().

Now we have these combinations:
socket   addr
IPv4     IPv4     use address "as is"
IPv4     IPv6     programming error, connect() will fail
IPv6     IPv4     use osiSockIPv4toIPv6()
IPv6     IPv6     use address "as is"
Changes to be committed:
	modified:   modules/ca/src/client/repeater.cpp
	modified:   modules/ca/src/client/tcpiiu.cpp
	modified:   modules/libcom/src/log/logClient.c
	modified:   modules/libcom/src/osi/osiSock.c
	modified:   modules/libcom/src/osi/osiSock.h