Use epicsBaseDebugLog() instead of osiDebugPrint()
Use the more portable version for debug-prints, which should compile
under older versions of MSVC as well.

As Michael Davidsaver points out, using stderr could be a choice.
However, this is how it is:
- Using stderr:
  Is unbuffered, does not slow down the system when too many
  logs are produced. And logs may get lost.
- Using stdout:
  Is typically buffered.
  Sending too many logs will slow donw the system

For the moment, we go for stdout.
In the long term, this debug log should be reduced.
Eventually even removed, but we are not there yet.