IPv6: cast_server: Ignore IPv6/IPv4 if not enabled
Ignore IPv6 (and/or) IPv4 messages, if specified.
In short, look at EPICS_CAS_AUTO_BEACON_ADDR_LIST and EPICS_CA_AUTO_ADDR_LIST
and determine what to do.
If one of those (or both) is set to '46', handle both IPv4 and IPv6
If both are set to '6', handle only IPv6 - ignore IPv4

Else, handle IPv4 and ignore IPv6.
Regardless if base is compiled with "#if EPICS_HAS_IPV6" or not