IPv6: epicsSocket46BindLocalPort() needs family
Make epicsSocket46BindLocalPort() slightly more versatile:
It gets one more parameter, the address-famliy of the socket.
In this way, both IPv4 and IPv6 socket can be bound to a local port number.
So far, "repeater registrations" are still in IPv4, so the
code doing the "bind-to-local-port" can be simplified.

In the long term, repeater regsitratons should be possible from an IPv6
address, which is probably "localhost anyway, so that the parameter
may go away again.