repeater: Do not loose a socket in register_new_client()
When a client registers with the repeater, a check is done that
the client is on the same host.
If a host has e.g. an ethernet interface with address 1.2.3.4
both registrations from 1.2.3.4 or 127.0.0.1 are accepted.
All others are ignored.
The check is done by creating a temporary socket and bind() it to
the IP address to be checked, 1.2.3.4 in the example, with port = 0

When this succeeds, the registration can continue.
The current code has possible improvements:
- The temporary socket is not closed (and kind of lost)
- When there are 2 or more ethernet interfaces, only one is handled.
  The one where the first registration is received.

Solution:
Close the socket after bind() and simplify the rest of the code