IPv6: Fix is-local-address check for new clients
Fix a bug introduced by a mis-understanding:
When a new client registers (via IPv4) and the sender address
is not 127.0.0.1, those registrations are ignored - unless they
come from the same host, and are using one of the local ethernet
interface addresses.
Trying to bind a test socket to this address will succeed then.