Ipv6: Allow repeater registrations via IPv6
Allow the repeater to accept registrations via IPv6.
The repeater itself opens a IPv6 socket and listens to it,
instead of the IPv4 socket.
It will receive both IPv4 "[::ffff:127.0.0.1]"
and IPv6 "[::1]" registrations.

Change udpiiu to send registration requests alternating
on both IPv6 and IPv4.
This seems a little bit hackish, but commit it, and improve it later.