IPv6: Print sockets as int, not unsigned long
Change the way we print sockets:
Under Unix it is an int, only windows uses a handle.
And a long is 32 bit. So casting a SOCKET to "long" or "unsigned long"
does not give any more inforamtion anyway.
So: cast a SOCKET into an int on all platforms and use "%d"