IPv6: Add osiSockPoll()
Many platforms (vxworks, RTEMS) do not have poll(),
for others they have poll(), but poll() was not reported to
be reliable.
To make things consistant, write a very limited osiSockPoll() that uses
select().

Implement only the "reading" part, we do not need more than that,
and it will be untested anyway.