IPv6: epicsSocket46Connect() takes `struct sockaddr *`
Change the connect() wrapper to a take struct sockaddr * and
addrlen as paramters for the remote to connect to.

Side note: The address family which created the socket is there as well.
In other words:
A socket created with AF_INET6 can connect to an AF_INET address, if needed.